Python port Phase 0: pybind11 engine binding with bit-exact R parity
Proves the core thesis of PYTHON_PORT.md -- the muse C++ engine can serve
a second front-end unchanged.  A pybind11 module wraps runMuseCommand()
exactly as musecpp2R.cpp wraps it for R.

- src/muse_compat.h: Rprintf shim, active only under -DMUSE_PYTHON_BUILD;
  the engine's sole R C-API dependency.  No-op under the R build.
- src/musecore.h: include the shim before PTSmodel.h.
- src/python/musecpp2py.cpp: pybind11 module `_musecore.ucomp`, a
  field-for-field mirror of the R marshalling (numpy<->Armadillo by copy;
  carma adoption deferred to Phase 1's CMake packaging).
- src/PTSmodel.h: one portability fix -- cast arma integer element product
  to int in a std::max call (system Armadillo's sword is long long; the
  R toolchain's matched int).  Numerically neutral; R golden master
  unchanged.
- python/: venv-based skeleton, direct build_spike.sh (CMake/scikit-build
  packaging is Phase 1), and the parity harness: dump_reference.R emits
  byte-identical engine inputs + R outputs, test_engine_parity.py feeds
  them to the binding and compares.

Result: 6/6 fixed-spec cases (incl. damped trend and ARMA(1,0)) match R
with worst abs diff 0.0e0 -- identical, since it is the same engine.
R build/tests and the 147-cell golden master remain green.

+++ src/PTSmodel.h
@@ -3162,9 +3162,11 @@ void BSMclass::initParBsm(){
                     }
                     int maxLag = 0;
                     for (uword b = 0; b < arOrders.n_elem; ++b)
-                        maxLag = std::max(maxLag, arOrders(b) * armaLags(b));
+                        maxLag = std::max(maxLag,
+                                          static_cast<int>(arOrders(b) * armaLags(b)));
                     for (uword b = 0; b < maOrders.n_elem; ++b)
-                        maxLag = std::max(maxLag, maOrders(b) * armaLags(b));
+                        maxLag = std::max(maxLag,
+                                          static_cast<int>(maOrders(b) * armaLags(b)));
                     arma::vec pacf = (maxLag > 0)
                         ? sampleYWpacf(SSmodel::inputs.y, maxLag)
                         : arma::vec();
